messages_service/requirements.txt
fastapi
uvicorn
pika
python-consul
